+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {BrowserRouter, Navigate, Route, Routes} from 'react-router-dom';
 import ROUTES from './constants/routes';
-import TodoListPage from './pages/TodoList';
+import TodoPage from './pages/Todo';
 
 function App() {
     return (
@@ -10,7 +10,7 @@ function App() {
                 <Route path='/' element={<Navigate to={ROUTES.TODO} />} />
                 <Route path={ROUTES.SIGNIN} />
                 <Route path={ROUTES.SIGNUP} />
-                <Route path={ROUTES.TODO} element={<TodoListPage />} />
+                <Route path={ROUTES.TODO} element={<TodoPage />} />
                 <Route path='/*' element={<Navigate to={ROUTES.TODO} />} />
             </Routes>
         </BrowserRouter>
