--- src/components/TodoList/List.tsx
@@ -1,9 +1,0 @@
-const TodoList = () => {
-    return (
-        <>
-            <h1>TodoList Component</h1>
-        </>
-    );
-};
-
-export default TodoList;
